+++ packages/api-bzz-base/src/index.js
@@ -77,7 +77,7 @@ export default class BaseBzz {
     this._url = new URL(url).toString()
   }
 
-  _getDownloadURL(
+  getDownloadURL(
     hash: string,
     options: DownloadOptions,
     raw?: boolean = false,
@@ -95,7 +95,7 @@ export default class BaseBzz {
     return url
   }
 
-  _getUploadURL(options: UploadOptions, raw?: boolean = false): string {
+  getUploadURL(options: UploadOptions, raw?: boolean = false): string {
     // Default URL to creation
     let url = this._url + BZZ_MODE_PROTOCOLS[raw ? 'raw' : 'default']
     // Manifest update if hash is provided
@@ -108,16 +108,22 @@ export default class BaseBzz {
     return url
   }
 
-  hash(domain: string): Promise<string> {
-    return this._fetch(`${this._url}bzz-hash:/${domain}`).then(resText)
+  hash(domain: string, headers?: Object = {}): Promise<string> {
+    return this._fetch(`${this._url}bzz-hash:/${domain}`, { headers }).then(
+      resText,
+    )
   }
 
-  list(hash: string, options?: DownloadOptions = {}): Promise<ListResult> {
+  list(
+    hash: string,
+    options?: DownloadOptions = {},
+    headers?: Object = {},
+  ): Promise<ListResult> {
     let url = `${this._url}bzz-list:/${hash}`
     if (options.path != null) {
       url += `/${options.path}`
     }
-    return this._fetch(url).then(resJSON)
+    return this._fetch(url, { headers }).then(resJSON)
   }
 
   _download(
@@ -125,12 +131,16 @@ export default class BaseBzz {
     options: DownloadOptions,
     headers?: Object = {},
   ): Promise<*> {
-    const url = this._getDownloadURL(hash, options)
+    const url = this.getDownloadURL(hash, options)
     return this._fetch(url, { headers }).then(resOrError)
   }
 
-  download(hash: string, options?: DownloadOptions = {}): Promise<*> {
-    return this._download(hash, options)
+  download(
+    hash: string,
+    options?: DownloadOptions = {},
+    headers?: Object,
+  ): Promise<*> {
+    return this._download(hash, options, headers)
   }
 
   _upload(
@@ -139,18 +149,19 @@ export default class BaseBzz {
     headers?: Object = {},
     raw?: boolean = false,
   ): Promise<string> {
-    const url = this._getUploadURL(options, raw)
+    const url = this.getUploadURL(options, raw)
     return this._fetch(url, { body, headers, method: 'POST' }).then(resText)
   }
 
   uploadFile(
     data: string | Buffer,
     options?: UploadOptions = {},
+    headers?: Object = {},
   ): Promise<string> {
     const body = typeof data === 'string' ? Buffer.from(data) : data
     const raw = options.contentType == null
-    const headers: Object = { 'content-length': body.length }
-    if (!raw) {
+    headers['content-length'] = body.length
+    if (headers['content-type'] == null && !raw) {
       headers['content-type'] = options.contentType
     }
     return this._upload(body, options, headers, raw)
@@ -159,6 +170,7 @@ export default class BaseBzz {
   uploadDirectory(
     _directory: DirectoryData,
     _options?: UploadOptions,
+    _headers?: Object,
   ): Promise<string> {
     return Promise.reject(new Error('Must be implemented in extending class'))
   }
@@ -166,15 +178,20 @@ export default class BaseBzz {
   upload(
     data: string | Buffer | DirectoryData,
     options?: UploadOptions = {},
+    headers?: Object = {},
   ): Promise<string> {
     return typeof data === 'string' || Buffer.isBuffer(data)
       ? // $FlowFixMe: Flow doesn't understand type refinement with Buffer check
-        this.uploadFile(data, options)
-      : this.uploadDirectory(data, options)
+        this.uploadFile(data, options, headers)
+      : this.uploadDirectory(data, options, headers)
   }
 
-  deleteResource(hash: string, path: string): Promise<string> {
-    const url = this._getUploadURL({ manifestHash: hash, path })
-    return this._fetch(url, { method: 'DELETE' }).then(resText)
+  deleteResource(
+    hash: string,
+    path: string,
+    headers?: Object = {},
+  ): Promise<string> {
+    const url = this.getUploadURL({ manifestHash: hash, path })
+    return this._fetch(url, { method: 'DELETE', headers }).then(resText)
   }
 }