+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fix `defaultPath` support for uploads in `api-bzz-browser` and `swarm-client`.
 - Add `--default-path` flag to `bzz:upload` CLI command.
 - Fix flow types related to `fs-extra` usage.
+- Add optional `headers` argument to `Bzz` class methods.
+- Expose `getDownloadURL()` and `getUploadURL()` utility methods in `Bzz` class.
 
 ## v0.4.0 (2018-09-19)